Move Cosmos3TransferSetupStep to before_encoder.py
